+++ src/bm25.cc
@@ -33,29 +33,25 @@ namespace fuzzy
     return sidx;
   }
 
-  double BM25::bm25_score_pattern(
+  float BM25::bm25_score_pattern(
     unsigned s_id,
     std::vector<unsigned> pattern_wids) const
   {
-    double score = 0;
+    float score = 0;
     for (unsigned& term : pattern_wids){
       score += (*_bm25)[term][s_id];
-      // std::cerr << term << ":" << (*_bm25)[term][s_id] << " | ";
     }
-    // std::cerr << std::endl;
     return score;
   }
 
-  double BM25::bm25_score(
+  float BM25::bm25_score(
     unsigned term,
     unsigned s_id,
-    double avg_doc_length,
+    float avg_doc_length,
     boost::multi_array<unsigned, 2>& tf,
     std::vector<float>& idf)
   {
     /* BM25 formula */
-    // std::cerr << idf[term] * (_k1 + 1) * tf[term][s_id] / (tf[term][s_id] + _k1 * ((1 - _b) + (_b * (get_sentence_length(s_id) / avg_doc_length)))) << " ";
-
     return idf[term] * (_k1 + 1) * tf[term][s_id] / 
           (tf[term][s_id] + _k1 * ((1 - _b) + (_b * (get_sentence_length(s_id) / avg_doc_length))));
   }
@@ -64,7 +60,6 @@ namespace fuzzy
     if (_sorted)
       return;
 
-    // std::cerr << "build TF + DF..." << std::endl;
     // build TF + DF
     boost::multi_array<unsigned, 2> tf(boost::extents[vocab_size][_sentence_pos.size()]);
     std::fill(tf.data(), tf.data() + tf.num_elements(), 0);
@@ -79,42 +74,31 @@ namespace fuzzy
         relative_pos = 0;
         sentence_idx++;
         seen_in_doc.clear();
-        // std::cerr << "| ";
       }
       else if (_sentence_buffer[pos] != 0)
       {
         tf[_sentence_buffer[pos]][sentence_idx]++; // term frequency count
         auto result_insert = seen_in_doc.insert(_sentence_buffer[pos]);
-        // std::cerr << "(" << result_insert.second << ") ";
         if (result_insert.second)
           doc_freq[_sentence_buffer[pos]]++;
-        // std::cerr << _sentence_buffer[pos] << " ";
       }
     }
-    // std::cerr << std::endl;
-    // for (auto &e : doc_freq)
-    //   std::cerr << e << " ";
-    // std::cerr << std::endl;
-    // std::cerr << "build IDF from DF..." << std::endl;
     // build IDF from DF
     std::vector<float> idf(vocab_size);
     for (unsigned term = 0; term < vocab_size; term++)
     {
       idf[term] = std::log(
-        ((double)_sentence_pos.size() - (double)doc_freq[term] + 0.5) /
-        ((double)doc_freq[term] + 0.5)
+        ((float)_sentence_pos.size() - (float)doc_freq[term] + 0.5) /
+        ((float)doc_freq[term] + 0.5)
       );
     }
-    // std::cerr << "build BM25 from TF and IDF..." << std::endl;
     // build BM25 from TF and IDF
-    double avg_doc_length = (double)_sentence_buffer.size() / (double)_sentence_pos.size() - 2;
-    // std::cerr << vocab_size << ", " << _sentence_pos.size() << std::endl;
+    float avg_doc_length = (float)_sentence_buffer.size() / (float)_sentence_pos.size() - 2;
     _bm25 = new boost::multi_array<float, 2>(boost::extents[vocab_size][_sentence_pos.size()]);
     for (unsigned s_id = 0; s_id < _sentence_pos.size(); s_id++)
       for (unsigned term = 0; term < vocab_size; term++)
         (*_bm25)[term][s_id] = bm25_score(term, s_id, avg_doc_length, tf, idf);
 
-    // std::cerr << "sorted" << std::endl;
     _sorted = true;
   }
 }
